Moving script and tests. Other recommended fixes
